requirements.txt
numpy==1.20.1
faiss-cpu==1.6.3
fastapi==0.62.0
uvicorn==0.11.3
pydantic==1.5.1
apscheduler==3.7.0
CMRESHandler==1.0.0
requests==2.26.0
absl-py
pyyaml
